simple convert to OOP
